+++ common/datasets/switchboard/__init__.py
@@ -1,4 +1,4 @@
-from i6_experiments.common.datasets.switchboard.eval import *
+from i6_experiments.common.datasets.switchboard.corpus_eval import *
 from i6_experiments.common.datasets.switchboard.export import export_all
 from i6_experiments.common.datasets.switchboard.lexicon import *
-from i6_experiments.common.datasets.switchboard.train import *
+from i6_experiments.common.datasets.switchboard.corpus_train import *
